backend/requirements.txt
# Python dependencies for the FastAPI backend service
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.0  # Required for TestClient

# ML and embedding dependencies for M2
sentence-transformers==2.2.2
numpy==1.24.3

# Configuration and ranking dependencies for M3
pyyaml==6.0.1

# LLM dependencies for M4
openai==1.12.0

# LTR dependencies for M7
scikit-learn==1.3.0
joblib==1.3.2
